refactor: move play_tic_tac_toe function to utils module
